split repository into user, bound and filter repositories
