+++ InstaBotPrototype/InstaBotLibrary/Repositories/IFilterRepository.cs
@@ -1,0 +1,14 @@
+﻿using InstaBotLibrary.Models;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace InstaBotLibrary.Repositories
+{
+    interface IFilterRepository
+    {
+        List<FilterModel> getBoundFilters(int boundId);
+        void AddFilter(FilterModel filter);
+        void DeleteFilter(FilterModel filter);
+    }
+}
